code/scripts/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
websockets>=13.0
langchain>=0.3.0
langchain-ollama>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
weaviate-client>=4.9.0
pypdf>=5.0.0
unstructured>=0.16.0
markdown>=3.7
pyyaml>=6.0
python-multipart>=0.0.12
